requirements.txt
asgiref==3.4.1
distlib==0.3.4
Django==3.2.8
django-filter==21.1
django-property-filter==1.1.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
filelock==3.5.0
geographiclib==1.52
geopy==2.2.0
Pillow==9.0.1
platformdirs==2.5.0
psycopg2==2.9.3
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
virtualenv==20.13.1
